drawCommand: make it a setuppable class
+ this whole unique_ptr spiel looked a lot like code-smell and an
inappropriate attempt to protect against Macciavelli, which you really
can't.

Draw Commands are now default objects, and use the compiler provided 5
constructors, and the compiler provided deleter. As this is a
non-owning, more or less pure data class this seems to be the best way
forward.

+++ libs/openFrameworks/vk/RenderContext.h
@@ -218,7 +218,7 @@ inline ::vk::CommandBuffer RenderContext::requestAndBeginPrimaryCommandBuffer(){
 			.setRenderPass( getRenderPass() )
 			.setFramebuffer(getFramebuffer() )
 			.setRenderArea( getRenderArea() )
-			.setClearValueCount( clearValues.size() )
+			.setClearValueCount( uint32_t(clearValues.size()) )
 			.setPClearValues( clearValues.data() )
 			;
